Move all OTLP exporter tests that are testing the underlying behavior in the mixin to the mixin unit tests, instead of having them specified multiple times in the Metric/Log/Trace exporters.
Fix the shutdown tests which were flaky, so that
they just test whether a pending export call completes
or not.  Update shutdown so it doesn't release the
lock -- in cases where an export call is pending,
export then also releases the lock causing a Runtime
Error: https://docs.python.org/3/library/threading.html#threading.Lock.release.
